+++ houssg/src/components/owner/management/ManageReview.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ManageReview = () => {
+	return <div>리뷰목록입니다</div>;
+};
+
+export default ManageReview;
