+++ houssg/src/components/owner/management/ManageRoomWrap.tsx
@@ -1,0 +1,5 @@
+const ManageRoomWrap = () => {
+	return <div>룸입니다</div>;
+};
+
+export default ManageRoomWrap;
