+++ houssg/src/components/owner/management/ManageTabComp.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
+import { ManageRoomWrap } from '.';
 
 const ManageTabComp: React.FC<{ isRoomSelected: boolean }> = ({ isRoomSelected }) => {
-	return <div>{isRoomSelected ? <div>룸</div> : <div>리뷰</div>}</div>;
+	return <div>{isRoomSelected ? <ManageRoomWrap /> : <div>리뷰</div>}</div>;
 };
 
 export default ManageTabComp;
