+++ houssg/src/components/owner/management/index.ts
@@ -1,3 +1,9 @@
 import ManageWrapComp from './ManageWrapComp';
 import ManageHouseRead from './ManageHouseRead';
-export { ManageWrapComp, ManageHouseRead };
+import ManageHouseEdit from './ManageHouseEdit';
+import ManageNav from './ManageNav';
+import ManageReview from './ManageReview';
+import ManageRoomWrap from './ManageRoomWrap';
+import ManageTabComp from './ManageTabComp';
+
+export { ManageWrapComp, ManageHouseRead, ManageHouseEdit, ManageNav, ManageReview, ManageRoomWrap, ManageTabComp };
