+++ houssg/src/types/manage.ts
@@ -2,4 +2,12 @@ interface ManageHouseProps {
 	setIsEditMode: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-export type { ManageHouseProps };
+interface ManageNavProps {
+	isRoomSelected: boolean;
+	setSelectedNav: React.Dispatch<React.SetStateAction<boolean>>;
+	isOpenTabComp: boolean;
+	setIsOpenTabComp: React.Dispatch<React.SetStateAction<boolean>>;
+	setIsEditMode: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+export type { ManageHouseProps, ManageNavProps };
